Add list of audio engines to "Buses" section when multiple audio engines exist
Note that this does not work well because the audio engine `uniqueId` gets used for 2 different tree items, once in the "Audio Engines" section and again in the "Buses" section. This goes against the main design and will mess with the view's state.

I'm thinking drop the "Audio Engines" section, rename combine the "Buses" and "Sounds" sections, with sounds in the bus hierarchy, and then always use the audio engines as the root entities for the unified "Sounds" section hierarchy.