Silence console error from unit test that is passing
The PBRMaterial unit test "setPrePassRenderer > should set prepass renderer when scattering enable" passes but causes the prePassRendererSceneComponent.ts module to emit an error to the JS console, which shows red in the output and makes it look like the test failed even though it passed.

This change uses the BabylonJS `Logger` class to silence the `console.error` for this test.